ICUI-11: Moved files around
**Problem:**

**Fix:**
TODO: put login and register buttons on the side?
**Testing:**
