+++ src/components/statistics/StatisticsPage.tsx
